requirements.txt
openai
datasets
notebook
scikit-learn
faiss-cpu
sentence-transformers
dotenv
gradio
fastapi
uvicorn
# langchain[openai]
# langchain-community
# langchain-core
# langchain_huggingface
# docling
